fix(stella-cli): --tools '*:off,…' advertises exactly the requested set
The issue's spec put 13 tools on the wire when 5 were requested, through
two independent defects:

Family 1 (grep/glob/graph_query/semantic_code_search): the catalog group
holding the search family was itself named 'search' — the same name as
the fused search tool (#3076) — so the spec's 'search:on' also resolved
as a group switch and re-enabled all four siblings past the wildcard
deny. The family is now grouped as 'retrieval'; a catalog invariant test
pins that a group may only share a tool's name when it holds exactly
that tool (the 'task' collision is the declared exemption, #3192).

Family 2 (tool_search/skill_search/mcp_search/invoke_skill):
DiscoveryToolSet is layered above PolicyToolSet, so its four session
tools were appended after the policy filter and executed without ever
consulting it. The discovery layer now carries the session's ToolPolicy
(DiscoveryToolSet::for_session, derived from the same
session_tool_policy every driver hands PolicyToolSet) and applies it to
its own additions on both sides — withheld from schemas() and refused
by execute() in the unknown-tool wording. The policy-free constructor
is cfg(test), so no production call site can omit the policy. Lean-mode
hiding stays permissive on execute, unchanged (#3100).

Witness (crates/stella-cli/src/tool_policy.rs): the real session stack
under the issue's exact spec advertises exactly the five names — 13 on
origin/main (matching the proxy capture name-for-name), 5 here — plus a
byte-identity test that --tools X and STELLA_LEAN_TOOLS=only:X produce
the same tools array, and a two-sided hidden-and-refused witness for
the discovery gate.

discovery.rs's inline test module moved verbatim to
discovery/tests.rs (file-size headroom); no test deleted.

Closes #3120
Refs #3192
