+++ crates/stella-cli/src/command_deck.rs
@@ -4231,7 +4231,7 @@ async fn run_lead_turn(
         // Discovery layer above the policy filter (it must see the full
         // *permitted* catalog), below the taps (searches are read-only; taps
         // watch writes).
-        let tools = crate::discovery::DiscoveryToolSet::new(&gated, cfg.workspace_root.clone())
+        let tools = crate::discovery::DiscoveryToolSet::for_session(&gated, cfg)
             .with_project_prompts_allowed(cfg.authority.project_prompts_allowed)
             .with_activation(activated.clone());
         let tapped = TaskTap {
