+++ website/content/docs/agent-tools/index.mdx
@@ -340,7 +340,7 @@ Every tool ships **on**, `bash` and the web family included. The one way to turn
 }
 ```
 
-A key is a **tool name**, a **group name** (the family headings above — `file`, `search`, `process`, `repo`, `web`, `bash`, `task`, …, plus `mcp` and `custom` for tools the built-in catalog has never heard of), or `"*"` for everything. Most specific wins: name beats group beats `"*"`, and anything unmentioned is on.
+A key is a **tool name**, a **group name** (the family headings above — `file`, `retrieval` (the Search family; deliberately not keyed `search`, so that key addresses the `search` tool alone), `process`, `repo`, `web`, `bash`, `task`, …, plus `mcp` and `custom` for tools the built-in catalog has never heard of), or `"*"` for everything. Most specific wins: name beats group beats `"*"`, and anything unmentioned is on.
 
 Two recipes that fall out of that rule:
